Point and Linestring geos scalar types, wrapped by traits
