+++ src/io/geos/array/mod.rs
@@ -1,2 +1,3 @@
 pub mod linestring;
+pub mod point;
 pub mod polygon;
