+++ src/io/geos/scalar/mod.rs
@@ -8,3 +8,6 @@ pub mod multipoint;
 pub mod multipolygon;
 pub mod point;
 pub mod polygon;
+
+pub use linestring::GEOSLineString;
+pub use point::GEOSPoint;
